+++ libexec/scoop-unhold.ps1
@@ -1,23 +1,43 @@
 # Usage: scoop unhold <app>
 # Summary: Unhold an app to enable updates
+# Help: To unhold a user-scoped app:
+#      scoop unhold <app>
+#
+# To unhold a global app:
+#      scoop unhold -g <app>
+#
+# Options:
+#   -g, --global  Unhold globally installed apps
 
+. "$PSScriptRoot\..\lib\getopt.ps1"
 . "$PSScriptRoot\..\lib\json.ps1" # 'save_install_info' (indirectly)
 . "$PSScriptRoot\..\lib\manifest.ps1" # 'install_info' 'Select-CurrentVersion' (indirectly)
 . "$PSScriptRoot\..\lib\versions.ps1" # 'Select-CurrentVersion'
 
-$apps = $args
+$opt, $apps, $err = getopt $args 'g' 'global'
+if ($err) { "scoop unhold: $err"; exit 1 }
 
-if(!$apps) {
+$global = $opt.g -or $opt.global
+
+if (!$apps) {
     my_usage
+    exit 1
+}
+
+if ($global -and !(is_admin)) {
+    error 'You need admin rights to unhold a global app.'
     exit 1
 }
 
 $apps | ForEach-Object {
     $app = $_
-    $global = installed $app $true
 
-    if (!(installed $app)) {
-        error "'$app' is not installed."
+    if (!(installed $app $global)) {
+        if ($global) {
+            error "'$app' is not installed globally."
+        } else {
+            error "'$app' is not installed."
+        }
         return
     }
 
@@ -29,7 +49,7 @@ $apps | ForEach-Object {
     $dir = versiondir $app $version $global
     $json = install_info $app $version $global
     $install = @{}
-    $json | Get-Member -MemberType Properties | ForEach-Object { $install.Add($_.Name, $json.($_.Name))}
+    $json | Get-Member -MemberType Properties | ForEach-Object { $install.Add($_.Name, $json.($_.Name)) }
     $install.hold = $null
     save_install_info $install $dir
     success "$app is no longer held and can be updated again."